requirements-dev.txt
-e .
flake8
pytest >= 3.0.0
pytest-django
pytest-cov
testtools
